+++ crates/px-configuration/src/lib.rs
@@ -1,7 +1,5 @@
-mod global;
 mod profile;
 mod project;
 
-// pub use global::*;
 pub use profile::*;
 pub use project::*;