+++ crates/px/src/commands/masters/create_bias.rs
@@ -1,14 +1,14 @@
 use crate::{ExitStatus, printer::Printer, utils::to_fits_ext};
 use anyhow::Result;
 use px_cli::CreateBiasMasterArgs;
-use px_conventions::profile::ProfilePath;
+use px_index::ProfileIndex;
 use px_pipeline::master_bias::CreateMasterBiasPipeline;
 use siril_sys::Builder;
 
 pub(crate) async fn create_master_bias(
     args: CreateBiasMasterArgs,
     printer: Printer,
-    profile_path: ProfilePath,
+    index: ProfileIndex,
 ) -> Result<ExitStatus> {
     // Guard to make sure the input folder exists first
     if !args.raw_folder.exists() {
@@ -23,7 +23,7 @@ pub(crate) async fn create_master_bias(
             printer.error("Output bias folder does not exist")?;
             return Ok(ExitStatus::Error);
         }
-        None => profile_path.bias.clone(),
+        None => index.profile.bias.clone(),
     };
 
     let master = CreateMasterBiasPipeline::builder()
@@ -38,10 +38,7 @@ pub(crate) async fn create_master_bias(
     // Pretty print the result
     printer.success(format!("Master BIAS stacking completed: {:?}", master))?;
 
-    // TODO: Add this new master to the px_profile.yaml config for later uses
-    let mut profile = profile_path.load_config()?;
-    let config = profile.add_master(master.into())?;
-    profile_path.save_config(&config)?;
+    index.register_master(master).await?;
 
     Ok(ExitStatus::Success)
 }