+++ crates/px/src/commands/masters/create_flat.rs
@@ -1,14 +1,14 @@
 use crate::{ExitStatus, printer::Printer, reporters::DefaultPipelineReporter, utils::to_fits_ext};
 use anyhow::Result;
 use px_cli::CreateFlatMasterArgs;
-use px_conventions::profile::ProfilePath;
+use px_index::ProfileIndex;
 use px_pipeline::master_flat::CreateMasterFlatPipeline;
 use siril_sys::Builder;
 
 pub(crate) async fn create_master_flat(
     args: CreateFlatMasterArgs,
     printer: Printer,
-    profile_path: ProfilePath,
+    index: ProfileIndex,
 ) -> Result<ExitStatus> {
     // Guard to make sure the input folder exists first
     if !args.raw_folder.exists() {
@@ -23,7 +23,7 @@ pub(crate) async fn create_master_flat(
             printer.error("Output flat folder does not exist")?;
             return Ok(ExitStatus::Error);
         }
-        None => profile_path.flat.clone(),
+        None => index.profile.flat.clone(),
     };
 
     // TODO: Make bias optional and find the best by default else use specified or error
@@ -48,10 +48,7 @@ pub(crate) async fn create_master_flat(
     // Pretty print the result
     printer.success(format!("Master FLAT stacking completed: {:?}", master))?;
 
-    // TODO: Add this new master to the px_profile.yaml config for later uses
-    let mut profile = profile_path.load_config()?;
-    let config = profile.add_master(master.into())?;
-    profile_path.save_config(&config)?;
+    index.register_master(master).await?;
 
     Ok(ExitStatus::Success)
 }